Create a PyPi package (#2)
* Refactor package to add CLI

* Update .gitignore

* Add missing sub-package

* Refactor readme
