+++ .gitignore
@@ -1,4 +1,8 @@
 .ipynb_checkpoints/
 __pycache__/
 .vscode
-
+dist/
+*.egg-info/
+.venv/
+*venv/
+.pypirc
