+++ pyproject.toml
@@ -2,20 +2,56 @@
 name = "cp2k-spm-tools"
 description = "CP2K Scanning Probe Microscopy tools."
 version = "1.4.4"
+authors = [
+    { name = "Kristjan Eimre", email = "kristjaneimre@gmail.com" },
+    { name = "Edoardo Baldi", email = "edoardo.baldi@empa.ch" },
+]
+readme = "README.md"
+license = { file = "LICENSE" }
+requires-python = ">=3.8"
+classifiers = [
+    "Programming Language :: Python :: 3",
+    "License :: OSI Approved :: MIT License",
+    "Operating System :: OS Independent",
+]
 
 dependencies = [
-  "numpy >= 1.22, < 2",
-  "scipy >= 1.10, < 2",
-  "ase >= 3.15, < 4",
-  "matplotlib >= 3.0, < 4",
-  "mpi4py > 3.1",
+    "numpy >= 1.22, < 2",
+    "scipy >= 1.10, < 2",
+    "ase >= 3.15, < 4",
+    "matplotlib >= 3.0, < 4",
+    "mpi4py > 3.1",
 ]
+
+[project.urls]
+"Homepage" = "https://github.com/nanotech-empa/cp2k-spm-tools"
+"Bug Tracker" = "https://github.com/nanotech-empa/cp2k-spm-tools/issues"
 
 [project.optional-dependencies]
 dev = ["ruff", "pre-commit"]
 
+[project.scripts]
+cp2k-bader-bond-order = "cp2k_spm_tools.cli.bader_bond_order:main"
+cp2k-crop-orbs-wfn = "cp2k_spm_tools.cli.crop_orbs_wfn:main"
+cp2k-cube-from-wfn = "cp2k_spm_tools.cli.cube_from_wfn:main"
+cp2k-cube-operations = "cp2k_spm_tools.cli.cube_operations:main"
+cp2k-cube-single-column = "cp2k_spm_tools.cli.cube_single_column:main"
+cp2k-cube-split = "cp2k_spm_tools.cli.cube_split:main"
+cp2k-hrstm-from-wfn = "cp2k_spm_tools.cli.hrstm_from_wfn:main"
+cp2k-overlap-from-wfn = "cp2k_spm_tools.cli.overlap_from_wfns:main"
+cp2k-stm-sts-wfn = "cp2k_spm_tools.cli.stm_sts_from_wfn:main"
+cp2k-stm-sts-plot = "cp2k_spm_tools.cli.stm_sts_plotter:main"
+
+[build-system]
+requires = ["setuptools>=61.0"]
+build-backend = "setuptools.build_meta"
+
 [tool.setuptools]
-packages = ["cp2k_spm_tools"]
+packages = [
+    "cp2k_spm_tools",
+    "cp2k_spm_tools.cli",
+    "cp2k_spm_tools.hrstm_tools",
+]
 
 [tool.ruff]
 line-length = 120
